Tool to dump OSP sensors.
Simple IIO tool to dump OSP sensors. Entirely self contained.

The Makefile assumes builds occur on an x86 host. It will attempt
to build both ARM and x86 binaries.

Usage:
    osp-iio [-l|-d|-c COUNT] sensorname ....

Sensorname is one or more IIO device names to dump.
If -c is specified, it will iterate for up to COUNT times before
exiting. A value of -1 (default) will iterate for ever.

If -d is specified, the IIO device(s) will be disabled afterwards.

If -l is specified, the IIO devices on the system will be listed
and exit.

Signed-off-by: Hunyue Yau <hy-git@hy-research.com>